Android: Refactored code
