+++ Android/Android/app/src/main/res/layout/activity_device_list.xml
@@ -45,7 +45,7 @@
         />
 
     <Button
-        android:id="@+id/button_scan"
+        android:id="@+id/btn_scan"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:text="@string/button_scan"
