+++ Android/Android/app/src/main/res/layout/device_name.xml
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <TextView xmlns:android="http://schemas.android.com/apk/res/android"
-          android:layout_width="match_parent"
-          android:layout_height="wrap_content"
-          android:padding="5dp"
-          android:textSize="18sp"
+    android:padding="5dp"
+    android:textSize="19sp"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+
     />
