+++ Android/Android/app/src/main/res/values/colors.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#3F51B5</color>
-    <color name="colorPrimaryDark">#303F9F</color>
-    <color name="colorAccent">#FF4081</color>
+    <color name="colorPrimary">#132FC2</color>
+    <color name="colorPrimaryDark">#1C2561</color>
+    <color name="colorAccent">#F7A4C0</color>
 </resources>
