+++ Android/Android/app/src/main/res/values/strings.xml
@@ -1,57 +1,46 @@
 <?xml version="1.0" encoding="utf-8"?>
-    <!-- Copyright (C) 2014 The Android Open Source Project
-
-         Licensed under the Apache License, Version 2.0 (the "License");
-         you may not use this file except in compliance with the License.
-         You may obtain a copy of the License at
-
-              http://www.apache.org/licenses/LICENSE-2.0
-
-         Unless required by applicable law or agreed to in writing, software
-         distributed under the License is distributed on an "AS IS" BASIS,
-         WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-         See the License for the specific language governing permissions and
-         limitations under the License.
-    -->
-
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
 
-    <string name="app_name">MDP Group1</string>
-    <string name="action_settings">Settings</string>
-    <string name="intro_message">intro_message</string>
-<!--  BluetoothChat -->
-<string name="send">Send</string>
-<string name="not_connected">You are not connected to a device</string>
-<string name="bt_not_enabled_leaving">Bluetooth was not enabled. Leaving Bluetooth Chat.</string>
-<string name="title_connecting">connecting...</string>
-<string name="title_connected_to">connected to <xliff:g id="device_name">%1$s</xliff:g></string>
-<string name="title_not_connected">not connected</string>
+<string name="app_name">MDP Group11</string>
+<string name="action_settings">Settings</string>
+<string name="intro_message">intro_message</string>
 
-<!--  DeviceListActivity -->
-<string name="scanning">scanning for devices...</string>
-<string name="select_device">select a device to connect</string>
-<string name="none_paired">No devices have been paired</string>
-<string name="none_found">No devices found</string>
+<!--  String for Bluetooth -->
+
+<string name="Send">Send</string>
+<string name="bt_not_connected">Not connected to a device!</string>
+<string name="bt_not_enabled">Bluetooth not enabled. Leaving Bluetooth Chat.</string>
+<string name="title_connecting_">Connecting...</string>
+<string name="title_connected_to">Connected to <xliff:g id="device_name">%1$s</xliff:g></string>
+<string name="title_not_connected">Not connected</string>
+
+<!--  String for DeviceListActivity -->
+
 <string name="title_paired_devices">Paired Devices</string>
 <string name="title_other_devices">Other Available Devices</string>
 <string name="button_scan">Scan for devices</string>
+<string name="scanning">Scanning for devices...</string>
+<string name="select_device">Connect to the selected device</string>
+<string name="no_paired">No paired device</string>
+<string name="none_found">No devices found</string>
 
-<!-- Options Menu -->
+<!-- String for Options Menu -->
+
+<string name="set_config_1">Set String 1 (C1)</string>
+<string name="set_config_2">Set String 2 (C2)</string>
+<string name="swipe_input_gesture">Enable Gesture Control</string>
+<string name="show_data_strings">Show Data Strings</string>
+<string name="show_bt_service">Show Bluetooth Comm</string>
+<string name="image_recognition_string">Image Recognition String</string>
+<string name="mdf_string_2_discovered_obstacles">MDF String 2: Discovered Obstacles</string>
+<string name="mdf_string_1_explored_squares">MDF String 1: Explored Squares</string>
 <string name="secure_connect">Connect to Device</string>
-<string name="insecure_connect">Connect a device - Insecure</string>
-<string name="discoverable">Make Discoverable</string>
-    <string name="set_waypoint">Set Waypoint</string>
-    <string name="remove_waypoint">Remove Waypoint</string>
-    <string name="set_robot_position">Set Robot Position</string>
-    <string name="auto_map">Auto Update Map</string>
-    <string name="update_map">Manual Update Map</string>
-    <string name="set_config1">Set String 1 (C1)</string>
-    <string name="set_config2">Set String 2 (C2)</string>
-    <string name="swipe_input">Enable Gesture Control</string>
-    <string name="show_data_strings">Show Data Strings</string>
-    <string name="show_bluetooth_service">Show Bluetooth Comm</string>
-    <string name="image_recognition_string">Image Recognition String</string>
-    <string name="mdf_string_2_discovered_obstacles">MDF String 2: Discovered Obstacles</string>
-    <string name="mdf_string_1_explored_squares">MDF String 1: Explored Squares</string>
+<string name="insecure_connect">Connect a device (Insecure)</string>
+<string name="mk_discoverable">Make Device Discoverable</string>
+<string name="set_wp">Set Waypoint</string>
+<string name="remove_wp">Remove Waypoint</string>
+<string name="set_rbt_pos">Set Robot Position</string>
+<string name="auto_update_map">Auto Update Map</string>
+<string name="update_map">Manual Update Map</string>
 
 </resources>
